task1/requirements.txt
TikTokApi==5.2.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
prometheus-client==0.19.0
schedule==1.2.1
backoff==2.2.1
nest-asyncio==1.5.8
playwright==1.40.0
python-json-logger==2.0.7
ruff==0.11.7